add more standards email to avoid spurious replies
RFC3834 section 5 documents the Auto-Submitted field pretty well. The
Precedence field is more controversial (in RFC 2076 at least) but is
commonly used in mailing lists. finally, Archive-At is documented
RFC5064 and refers to an archived form of the email message, which is
a good match for the RSS <link> tag