multipart HTML support
with this change, by default, feeds get sent with the text/html
mimetype, which is more appropriate than before - as we were sending
HTML with the text/plain mimetype. this, unfortunately, means that we
strip out the link included in the beginning of the document, but that
was invalid HTML anyways (more or less - the link was unclickable, at
least) and may have confused poor email readers. at least the link is
now available in the Archive-At header.

furthermore, if the html2text filter is enabled (or, more precisely, a
"summary_plain" field is available in the feed entry), a text/plain
option is added *before* the text/html entry, which renders nicely in
Mutt in my tests, but should also render well in html-capable mail
readers (untested).