partial multipart support
we now support adding multiple parts into mailboxes, and parse each
one individually into plaintext.

this also makes the html2text plugin default, because it is such a
common feature. the alternative is to force everyone that needs it to
configure it by hand, which is annoying *and* it keeps people from
using *other* possible filter plugins.