add some limited parallelism tests
we can only test the output as the variables are set in subprocesses.

we also need to switch from get() to wait() in the subprocess API
because it seems the test suite raises issues that were not found
during manual testing, namely that parse() can return unserializable
errors from feedparser which crash the multiprocessing module when
it's trying to pass that back up. since we discard those results in
non-threaded parse() anyways, there's no reason why we should do this
here.