re-enable the accessing of results from multiprocessing
turns out there *is* a reason for using .get: it will raise exceptions
properly. i believe that .wait() may not return exceptions the same
way parse() would do. i now remember i explicitly used .get() for this
reason. maybe .wait() does as well, but better be safe than sorry
later.

to avoid crashing on weird feeds, we massage the bozo_exceptions
returned normally by the parse() function, which are *not* the same as
exceptions returned by parse() in abnormal situations. this is
necessary because the multiprocessing serializer cannot pass the
bozo_exceptions as is. besides, we only need its string representation (and even that...) so just return that.