return bytes properly in fetch
this fixes double-encoding issues detected in the previous test.

our documentation was *saying* we were returning bytes from fetch()
but, actually, we were returning a string. it *seems* that feedparser
expects bytes because it returns us an decoded string which, when
re-encoded (e.g. in email.make_message), is double-encoded when it was
originally a string. when we pass bytes, the output is correct.