refactor feeds to remove global
we now use a class static session that can be used from plugins

this slightly simplifies the implementation of plugins as we do not
need to load the module at a specific point to get the right
session. it also makes sense to regroup those functions in a class,
although we haven't put parse() in there yet.

this requires regenerating the planet cassette as it changed *and* we
didn't have a copy, because there was a bug in the integration unit
test: the betamax session wasn't used because it wasn't passed. it's
likely this bug also affected the VCR code as well.