API refactoring: merge classes in a controller
The FeedStorage and FeedFetcher class were fairly similar: they were
wrappers around the storage subsystem, which is basically a "model",
so they are a "controller" in the classic sense. So merge them both in
a single "FeedManager" which is a controller class that manipulates
the underlying object.

This is the first step of a larger refactoring: we may want to create
a Feed class of objects to manipulate feeds directly, for example
allowing filters to operate on OPML import/exports...