API refactoring: create a Feed class
A "Feed" is the basic representation of a RSS/Atom feed. It derives
from the feedparser dict but we use our class to regroup functionality
that was loose around the module, like normalize() and parse(). It
makes it easier (and possible!) to pass the feed object to the
multiprocessing module without casting it to a dict(), which means
plugins actually get access to the full Feed object.

This also has the advantage of removing the ambiguity of what a "feed"
is internally: it is explicitly the merge of the config declared by
the user and the metadata obtained from the feed.