add transmission plugin
this plugin is introduced because the exec plugin cannot safely guess
a subdirectory to save feed into without resorting to using
plugins. now the transmission plugin has nicer defaults and properly
guesses a default path based on the item title, which can be
appropriately modified by filter plugins as needed.