allow shorter plugin names on the commandline
The full plugin load path is quite a mouthful. Make it easier for our
users and allow them to use only the shorter module name, relative to
the `feed2exec.plugins` path, on the commandline only.

We keep the configuration file and internal use with the full plugin
path. This is to avoid double-resolving the path and ensures the
configuration file has a reliable plugin name.