refactor temp db use to simplify test suite
This implies that the *same* test database and configurations are
reused over the course of the various tests. This means we need to
cleanup or remove the database from time to time.

This is particularly nasty with the database because the SqliteStorage
backend freaks out when a database disappeares and it's already
connected to it, probably because the file descriptor changes or
something.