fix OPML import crash with missing config file
Somehow it's possible that load_first_config returns None, exactly in
the case when there is no pre-existing configuration file. We
duct-tape a fix that looks up in the default directory, but this is
impossible to unit test because the class variable is instanciated
before we can change the environment.

Closes: #1