split code of conduct and contribution guide
While they do relate to each other, one is somewhat boilerplate and
people may have read it before, while the other is specific to this
project.

Furthermore, some sites like GitHub expect to see a CODE_OF_CONDUCT
file in the root directory. It will also show source-using jedis the
existing of the code clearly when reading the source.

We keep a link from the contribution guide to make sure people are
aware of the existence of the guide as well. It is also a first-class
citizen in the table of contents now.