factor out project-specific links in a single document
We favor the include over a `rst_prolog` configuration because the
latter "hides" the fact that we need an include at all, which may mean
it will break on PyPI or plain rst2html used on GitLab or GitHub sites.

This makes the documentation more portable to other projects which can
easily customize those links in one unified place. Sphinx checks will
make sure that the internal references are correct and will warn if
pages like `contact` are missing.