fix planet test failure on older html2text
The test suite was failing on my laptop for some mysterious
reason. After investigation, it turns out I still had an older version
of html2text there which triggered a change in the output. We had that
covered in test_plugins, but not in test_main, so copy the excepted
failure fixture there as well.