add failing test for issue #5
We do not need to see the output of that feed, actually: we just want
to parse it. Therefore, make sure we can test any .xml file provided
even if it does not have a matching .mbx.

It is a little strange we test parsing of all feeds in the mbx plugin,
but since we do it anyways, I don't see why it should be moved out of
there.