fix parsing of empty links
For some obscure reason, with the test feed provided in #5, the
resulting link is no available in the feed at all, which crashes the
normalize routine. That is because urlparse.urlsplit returns a
different type on strings (which are used in the normal path) or
"None" (which is used when the link is not present).

We fix this by defaulting to the empty string, although this probably
means that feed2exec will have trouble adapting to that feed in the
first place, but that's another bug.

Closes: #5