warn instead of crashing on invalid dates
We used to completely crash when a feed had invalid or missing
dates. After reviewing the standards, it turns out this is not quite
valid behavior: RSS 0.90 and 0.91, for example, do not have dates at
all. Yet it seems to me a valid feed should minimally include *some*
timestamps and the more likely explanation for a missing "parsed"
field is that feedparser wasn't able to parse the feed properly

Therefore, turn this into a warning. This will be annoying as hell for
some users and feeds, unfortunately, but I don't think silently
ignoring those errors will be much better, as we *do* need a
timestamp (for example to generate valid emails) internally.

We fallback to the current time, for lack of a better alternative.

Closes: #7