add JSON output plugin
This plugin is very simple, if not trivial: it simply dumps all the
feed items in a JSON stream. This can be parsed by `jq` on the
commandline to diagnose feed problems, do scripting or whatever.

This is different from how plugins normally receive "arguments", which
are "shell-quoted". For example, the `exec` plugin, when called with
`{item.url} {item.title}` would receive it as a shell-quoted ("$URL"
"$TITLE") string, and not JSON.