switch to dateparser for PyPI and tests as well
Because the Debian package recommends dateparser, I had different
results running tests natively on Debian and within tox. This
harmonizes things and makes use of dateparser everywhere, warts and
everything.

It does fix a few dates in the test suite that were mistakenly marked
as succeeding, but were in fact failing to parse.