reduce noise level of 'missing time' problems
It seems like previous versions of feedparser would never trigger that
problem, and would fill in the date instead. Now I have feeds that
have this on *every* item and it generates a lot of noise in my cron
jobs.

Keep that lower, to the info level, so it doesn't trigger noise for
what is, after all, a fairly mild problem.