fix broken cache adapter support
We did not need to pass the if-modified-since header. All that was
needd was that we lookup (and return!) the cache value properly. So
also remove that from the database.

The way things were setup, the `get` parameter was returning something
CacheControl was not expecting (a tuple) instead of the proper
thing (a single blob). Our `value` lookup was also failing, and was
irrelevant.

We also don't need to keep the adapter around, it's already present in
the session's adapters member.