forcibly preset the builtin feed session
Without this setting, the wrong session gets initialized in the new
Feed object. Before the caching layer was implemented, this didn't
matter much because those sessions were never called.

But since we did that, the Feed() constructor will usually open and
therefore create the database if the session isn't already set. So set
it to some dummy value that will never work to keep the constructor
from setting it.