move session and fetching to the feed manager
Having the session and the network code in the "model" makes no sense:
that stuff belongs in the "controller".

Having it there made it particularly difficult to implement the
caching layer, as I had to pass the db_path around quite a bit.

Now we still have this insanity going on, because I just copy-pasted
code around until things (mostly worked), but this will clear the way
for more, and cleaner refactoring.

I will hopefully be able to support other parsers through such
refactoring as well.