remove class-level sticky session parameter
This cleans up a lot of stuff. Now we can treat the session as a
normal feed_manager parameter. Since there is usually only one
feed_manager in operation at any time, it is basically a static
member. And if it isn't, that's fine too: anyways the underlying
database structures do the right thing here and take care of locking
and everything.

This allows us to merge the betamax fixture into the feed_manager one,
which is necessary to get a (properly setup) session to do HTTP
requests.

This removes the new sessionCache() hack that was required to pass the
right db_path to the session configuration, which was the main goal of
this refactoring.