DRY: refactor common requirements out of tox and setup.py
Requirements are now stored only in setup.py. Before, we were telling
tox to delegate the test runs to setup.py, but this made it impossible
to pass arguments to pytest. By calling pytest directly, we ensure we
can do stuff like `-v` to see which test is failing, and so on.

We do this by storing the dependencies list for tests in the
extras_requires section of the dict, and reuse that in the
requirements file.

It also seems we had three lists of dependencies:

 setup_requires: setuptools_scm, pytest-runner, pytest-cov, sphinx
 extras_requires["dev"]: pytest, tox, pyflakes
 tests_require: pytest, betamax

We converge all of those on the latter, and remove the pytest-specific
dependencies from setup_requires. Presumably this will still work in
package builds, but it's possible the test_requires will have to be
merged into the setup_requires for that to work.

While we're here, we remove the fakeroot -u references. I never use
this and don't know where that's coming from.