fix default log level value
The patch in !8 made it so the log level was ERROR by default, which
is what the author fundamentally wanted, but it's not actually what
the behavior originally was.

Restore the previous behavior by adding another flag, --warning,
hidden by default, which will provide the default value for the
loglevel.

Urgh.