leave Freenode for oftc.net
This is becoming a long story, but basically Freenode was hijacked,
policies were changed to allow racism and bigotry, trolls have been
promoted to opers, and opers are hijacking channels. It's a shame,
really, but nothing we can do it seems:

https://gist.github.com/joepie91/df80d8d36cd9d1bde46ba018af497409

Moving to OFTC since that's where most of my other channels (Debian,
Tor) are.