Add support for saving page resources to the Wayback Machine too
The SPN2 API used by this is not well documented and does
not return proper status codes or easily usable error text,
so make it enabled by default but add an option to disable it
and disable it in the tests, which need the right return codes.

The SPN1 API does return more specific error text in HTML in
the same way as SPN2 in some cases, so process that too.

See-also: https://blog.archive.org/2019/10/23/the-wayback-machines-save-page-now-is-new-and-improved/