fix another test failure in newer feedparser
It seems like feedparser found a new way to default certain dates. We
were previously correctly parsing that date before, but for some
obscure reason feedparser is now defaulting to the feed's
updated_parsed field instead.

Typically, this was being parsed by dateparser, but it seems this is
either failing (which I cannot confirm locally) or ignored by
feedparser:

    anarcat@angela:feed2exec(fp6)$ python3
    Python 3.9.2 (default, Feb 28 2021, 17:03:44)
    [GCC 10.2.1 20210110] on linux
    Type "help", "copyright", "credits" or "license" for more information.
    >>> from dateparser import parse, __version__
    >>> parse("Sun, 15 Feb 2015 00:00:00")
    datetime.datetime(2015, 2, 15, 0, 0)
    >>> __version__
    '1.0.0'

It's not clear why this is failing, therefore. The result is wrong,
but I want tests to pass, so let's just get this over with.