try to get rid of setup.py in favor of pyproject.toml
There are a few things still failing:

 1. we do not know how to provide "packages", "package_data",
    "setup_requires" and "data_files"

 2. it's unclear how setuptools_scm works with new build systems, if
    at all

For now we just play it dumb and added "setuptools_scm" into the
dependencies list, but it probably doesn't really work. We also added
"packages" and "package_data" as poetry defines it (with "packages"
and "includes"), but it's not clear this would work with other build
systems.

As for data_files, it has been deprecated by setuptools, so it
should probably go away:

    https://setuptools.pypa.io/en/latest/references/keywords.html

click is not particularly helpful on how to fix this either...

https://click.palletsprojects.com/en/7.x/setuptools/#setuptools-integration

Also note that this drops Python 2 support, but we probably already
did that previously, if only when we added type hints.