add failing test and fix for cachecontrol 0.12.10 new API
This is a little more than the hotfix provided by @terceiro (which was
correct), but I figured I would add a unit test as well.

The original reporter's URL has stopped triggering the bug too,
presumably because they removed their Expires headers. Thankfully, I
had one of those lying around, so I made it work with an existing one.

I didn't bother creating a cassette for this, because I was worried
about creating more copyright issues (#21) with the sample URL. Now
that I use my own site, it seems we could use a cassette, but then I
ran out of time, and figured it would be a nice thing to have
networked tests, as a first step towards fixing #21.

/cc @pabs

I also looked at using "responses", but it seems to hijack the adapter
code path, which cachecontrol also does, so I predicted it would be
hell. This is another reason why this might have failed with
cassettes, now that I think of it.

Closes: #22