Draft: start implementing expires
Still a bunch of stuff missing, most notably actually expiring things,
but we're at least recording *something* to the database.

Note that we also add support for expiring feed items, which I believe
cachecontrol has nothing to do with, so that's something we should
handle separately.

So still more work to be done here to:

 * actually purge old records... somewhere?
 * handle item caching expiry
 * seriously consider sqlalchemy, because this is getting unwieldy
 * maybe split the expiry check in a separate function to cleanup
   set()
 * consider splitting out that code so that we can publish our own
   cachecontrol-sqlite?

We don't actually know what cachecontrol gives us as the `expires`
argument, need to audit the upstream source to figure that out as
well.

In short: just a first stab, incomplete.

See #23