use pytest.mark.network instead of xfail
This is a slightly less... bespoke approach: instead of going through
the environment, we pass the right argument to skip certain markers
from pytest.

The main downside is we just completely skip the test instead of
expecting its failure. But then it does more explicitly what we say we
did, which was "USENETWORK=no"...