support for multiple filter plugins
Right now, this only calls all filters with the same arguments. It
also only applies to filters, not output.

This is necessary for me because I wanted a "unique" filter (coming
up) on top of the ikiwiki one. This is not possible without seriously
mangling the ikiwiki filter itself: it would be *necessarily* unique,
and I don't actually want it to uniquify everything all the time!

This is basically untested.

I'm also amazed that mypy just let me do this. Obviously not enough
coverage here.